BACKEND/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
idna==3.2
login==0.0.6
mysqlclient==2.0.3
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
serializers==0.2.4
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
